+++ hw-7/src/main/java/ru/dankoy/hw5/core/repository/book/BookDao.java
@@ -1,4 +1,4 @@
-package ru.dankoy.hw5.core.dao.book;
+package ru.dankoy.hw5.core.repository.book;
 
 import java.util.List;
 import java.util.Optional;
