+++ hw-7/src/main/java/ru/dankoy/hw5/core/repository/genre/GenreDao.java
@@ -1,4 +1,4 @@
-package ru.dankoy.hw5.core.dao.genre;
+package ru.dankoy.hw5.core.repository.genre;
 
 import java.util.List;
 import java.util.Optional;
