+++ hw-7/src/main/java/ru/dankoy/hw5/core/service/commentary/CommentaryServiceHibernate.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.dankoy.hw5.core.dao.commentary.CommentaryDao;
+import ru.dankoy.hw5.core.repository.commentary.CommentaryDao;
 import ru.dankoy.hw5.core.domain.Book;
 import ru.dankoy.hw5.core.domain.Commentary;
 import ru.dankoy.hw5.core.exceptions.EntityNotFoundException;
