+++ hw-7/src/main/java/ru/dankoy/hw5/core/service/genre/GenreServiceHibernate.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.dankoy.hw5.core.dao.genre.GenreDao;
+import ru.dankoy.hw5.core.repository.genre.GenreDao;
 import ru.dankoy.hw5.core.domain.Genre;
 import ru.dankoy.hw5.core.exceptions.EntityNotFoundException;
