+++ hw-7/src/test/java/ru/dankoy/hw5/Hw5ApplicationDefaultTests.java
@@ -12,9 +12,9 @@ import org.springframework.context.ApplicationContext;
 import ru.dankoy.hw5.core.commands.AuthorCommand;
 import ru.dankoy.hw5.core.commands.BookCommand;
 import ru.dankoy.hw5.core.commands.GenreCommand;
-import ru.dankoy.hw5.core.dao.author.AuthorDaoHibernate;
-import ru.dankoy.hw5.core.dao.book.BookDaoHibernate;
-import ru.dankoy.hw5.core.dao.genre.GenreDaoHibernate;
+import ru.dankoy.hw5.core.repository.author.AuthorDaoHibernate;
+import ru.dankoy.hw5.core.repository.book.BookDaoHibernate;
+import ru.dankoy.hw5.core.repository.genre.GenreDaoHibernate;
 import ru.dankoy.hw5.core.service.author.AuthorServiceHibernate;
 import ru.dankoy.hw5.core.service.book.BookServiceHibernate;
 import ru.dankoy.hw5.core.service.genre.GenreServiceHibernate;
