+++ hw-7/src/test/java/ru/dankoy/hw5/core/repository/commentary/CommentaryDaoHibernateTest.java
@@ -1,4 +1,4 @@
-package ru.dankoy.hw5.core.dao.commentary;
+package ru.dankoy.hw5.core.repository.commentary;
 
 
 import static org.assertj.core.api.Assertions.assertThat;
