+++ hw-7/src/test/java/ru/dankoy/hw5/core/repository/genre/GenreDaoHibernateTest.java
@@ -1,4 +1,4 @@
-package ru.dankoy.hw5.core.dao.genre;
+package ru.dankoy.hw5.core.repository.genre;
 
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
