+++ hw-7/src/test/java/ru/dankoy/hw5/core/service/author/AuthorServiceJdbcTest.java
@@ -18,7 +18,7 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import ru.dankoy.hw5.core.dao.author.AuthorDao;
+import ru.dankoy.hw5.core.repository.author.AuthorDao;
 import ru.dankoy.hw5.core.domain.Author;
 import ru.dankoy.hw5.core.exceptions.EntityNotFoundException;
