+++ hw-7/src/test/java/ru/dankoy/hw5/core/service/book/BookServiceHibernateTest.java
@@ -20,9 +20,9 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import ru.dankoy.hw5.core.dao.author.AuthorDaoHibernate;
-import ru.dankoy.hw5.core.dao.book.BookDao;
-import ru.dankoy.hw5.core.dao.genre.GenreDaoHibernate;
+import ru.dankoy.hw5.core.repository.author.AuthorDaoHibernate;
+import ru.dankoy.hw5.core.repository.book.BookDao;
+import ru.dankoy.hw5.core.repository.genre.GenreDaoHibernate;
 import ru.dankoy.hw5.core.domain.Author;
 import ru.dankoy.hw5.core.domain.Book;
 import ru.dankoy.hw5.core.domain.Commentary;
