+++ hw-7/src/test/java/ru/dankoy/hw5/core/service/commentary/CommentaryServiceHibernateTest.java
@@ -20,10 +20,10 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import ru.dankoy.hw5.core.dao.author.AuthorDaoHibernate;
-import ru.dankoy.hw5.core.dao.book.BookDaoHibernate;
-import ru.dankoy.hw5.core.dao.commentary.CommentaryDao;
-import ru.dankoy.hw5.core.dao.genre.GenreDaoHibernate;
+import ru.dankoy.hw5.core.repository.author.AuthorDaoHibernate;
+import ru.dankoy.hw5.core.repository.book.BookDaoHibernate;
+import ru.dankoy.hw5.core.repository.commentary.CommentaryDao;
+import ru.dankoy.hw5.core.repository.genre.GenreDaoHibernate;
 import ru.dankoy.hw5.core.domain.Book;
 import ru.dankoy.hw5.core.domain.Commentary;
 import ru.dankoy.hw5.core.exceptions.EntityNotFoundException;
